chore: replace fast-json-stable-stringify with safe-stable-stringify
This is a performance improvement as the latter is actually faster.
The circular reference detection is a breaking change due to now
producing valid JSON. The former implementation caused the output to
be invalid JSON in such cases.

BigInt are from now on also stringified and might as such also be
breaking.

fast-json-stable-stringify x 18,765 ops/sec ±0.71% (94 runs sampled)
safe-stable-stringify x 30,367 ops/sec ±0.39% (96 runs sampled)